all testing code now in tests files and all files in one place
REXI is now a function rather than a class. It calls the function b_coeffcients and uses the class REXIParameters which is where L, mu and a are stored (these don't depend on h or M)

To approximate functions other than exp(ix), we just need to change b_coefficients (I think).